+++ src/components/Containers/Dashboard/Dashboard/Dashboard.js
@@ -4,66 +4,21 @@ import SearchCard from "components/UI/SearchCard/";
 import TeamMembersOverview from "components/Sections/TeamMembers/List/Overview";
 import TeamMembersTab from "components/Sections/TeamMembers/Tab";
 import TrainingSeriesOverview from "components/Sections/TrainingSeries/List/Overview";
-import TrainingSeriesTab from "components/Sections/TrainingSeries/TabTSComponents/TrainingSeriesTab.js";
+import TrainingSeriesTab from "components/Sections/TrainingSeries/Tabs/TrainingSeriesTab.js";
 import NotificationsCard from "components/Sections/Notifications/Card";
 import NotificationsOverview from "components/Sections/Notifications/Card/Overview/Overview.js";
 import Responses from "components/Sections/Notifications/Responses";
-
-import BottomNavigation from "@material-ui/core/BottomNavigation";
-import BottomNavigationAction from "@material-ui/core/BottomNavigationAction";
-import PermIdentity from "@material-ui/icons/PermIdentityOutlined";
-import Home from "@material-ui/icons/HomeOutlined";
-import QuestionAnswer from "@material-ui/icons/QuestionAnswerOutlined";
-import Timeline from "@material-ui/icons/TimelineOutlined";
-import ModeComment from "@material-ui/icons/ModeCommentOutlined";
+import TabNavigation from "./helpers/TabNavigation.js";
 
 import { TripleColumn, SmallColumns, DashWrapper } from "./styles.js";
 
 function Dashboard(props) {
   const [topTab, setTopTab] = useState("overview");
-  const { user_id } = props;
+  const { user_id, history } = props;
 
   return (
     <DashWrapper>
-      <BottomNavigation
-        value={topTab}
-        onChange={(e, value) => {
-          setTopTab(value);
-        }}
-        style={{
-          width: "100%",
-          whiteSpace: "nowrap",
-          background: "rgb(240,240,240)"
-        }}
-      >
-        <BottomNavigationAction
-          label="Overview"
-          value="overview"
-          icon={<Home />}
-          color="primary"
-        />
-        <BottomNavigationAction
-          label="Team Members"
-          value="team members"
-          icon={<PermIdentity />}
-        />
-        <BottomNavigationAction
-          label="Training Series"
-          value="training series"
-          icon={<Timeline />}
-        />
-        <BottomNavigationAction
-          label="Messages"
-          value="messages"
-          icon={<QuestionAnswer />}
-        />
-        <BottomNavigationAction
-          label="Responses"
-          value="responses"
-          icon={<ModeComment />}
-        />
-      </BottomNavigation>
-
+      <TabNavigation topTab={topTab} setTopTab={setTopTab} />
       {topTab === "overview" && (
         <TripleColumn>
           <SmallColumns>
@@ -73,16 +28,14 @@ function Dashboard(props) {
               containerTourNum="1"
               section="Team Members"
               headerTourNum={["2", "3"]}
-              handleAdd={() => props.history.push("/home/create-team-member")}
+              handleAdd={() => history.push("/home/create-team-member")}
             />
             <SearchCard
               user_id={user_id}
               List={TrainingSeriesOverview}
               containerTourNum="4"
               section="Training Series"
-              handleAdd={() =>
-                props.history.push("/home/create-training-series")
-              }
+              handleAdd={() => history.push("/home/create-training-series")}
             />
           </SmallColumns>
           <NotificationsCard
