+++ src/components/Sections/TeamMembers/Assign/styles.js
@@ -1,6 +1,26 @@
 import styled from "styled-components";
 
 export const styles = theme => ({
+  paper: {
+    width: "89%",
+    backgroundColor: theme.palette.background.paper,
+    boxShadow: theme.shadows[5],
+    padding: theme.spacing.unit * 4,
+    outline: "none",
+    margin: "5px auto",
+
+    "@media (max-width: 768px)": {
+      width: "89%",
+      // padding: 0,
+      margin: "5px auto"
+    },
+
+    "@media (max-width: 480px)": {
+      width: "80%",
+      // padding: 0,
+      margin: "5px auto"
+    }
+  },
   assignButton: {
     "margin-left": theme.spacing.unit,
     background: "#451476",
@@ -9,7 +29,6 @@ export const styles = theme => ({
       background: "#591a99",
       color: "white"
     },
-
     "@media (max-width: 768px)": {
       margin: "10px 5px"
     },
@@ -45,6 +64,14 @@ export const HeaderContainer = styled.div`
     padding: 10px 0;
     align-items: center;
     margin: 0 auto;
+    text-align: center;
+  }
+`;
+
+export const HolderText = styled.div`
+  margin: 50px 0;
+  p {
+    color: lightgray;
     text-align: center;
   }
 `;
