+++ src/components/Sections/TrainingSeries/Tabs/TrainingSeriesTabStyles.js
